Reduced the integer size for StepSize

+++ src/common/selection_list.rs
@@ -10,7 +10,7 @@ pub(crate) enum StepSize {
 }
 
 impl StepSize {
-    const fn amount(&self) -> usize {
+    const fn amount(&self) -> u8 {
         if matches!(self, Self::Many) {
             10
         } else {
